download is set up
