--- CoMoorValley.ClientWebServer/bin/Debug/netcoreapp1.1/ClientWebServer.runtimeconfig.dev.json
@@ -1,8 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\Kristen\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\Kristen\\.nuget\\packages"
-    ]
-  }
-}
