--- CoMoorValley.ClientWebServer/bin/Debug/netcoreapp1.1/ClientWebServer.runtimeconfig.json
@@ -1,12 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp1.1",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "1.1.2"
-    },
-    "configProperties": {
-      "System.GC.Server": true
-    }
-  }
-}
