--- CoMoorValley.ClientWebServer/static/gallery.js
@@ -1,37 +1,0 @@
-export const galleryPage = {
-  title: 'Photo Gallery',
-  index: null,
-  images: [
-    {
-      src: require('./..//src/assets/001.png'),
-      title: ''
-    }, {
-      src: require('./../src/assets/002.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/004.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/005.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/006.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/007.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/008.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/009.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/010.jpg'),
-      title: ''
-    }, {
-      src: require('./../src/assets/011.jpg'),
-      title: ''
-    }
-  ]
-}
