+++ CoMoorValley.ClientWebServer/static/pageContent/about.js
@@ -2,12 +2,12 @@ export const aboutPage = {
   title: 'About',
   infoRows: [
     {
-      src: require('./../src/assets/019.jpg'),
+      src: require('./../../src/assets/019.jpg'),
       content: 'Content topics here: Offerings such as pool, fitness center, etc? City info.',
       title: 'About CoMoor Valley',
       justification: 'left'
     }, {
-      src: require('./../src/assets/060.jpg'),
+      src: require('./../../src/assets/060.jpg'),
       content: 'Content topics here: When was it founded? Home many homes? Different types? Difference from CoMoor?',
       title: 'History',
       justification: 'right'
