+++ CoMoorValley.ClientWebServer/static/pageContent/home.js
@@ -1,4 +1,4 @@
 export const homePage = {
-  coMoorLogo: require('./../src/assets/logo_text.png'),
+  coMoorLogo: require('./../../src/assets/logo_text.png'),
   subtitle: 'Where the convenience of the city meets the beauty of the woods.'
 }
