+++ CoMoorValley.ClientWebServer/static/pageContent/news.js
